% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiEmpCdf.R
\name{BiEmpCdf}
\alias{BiEmpCdf}
\title{Empirical bivariate cdf}
\usage{
BiEmpCdf(data, y1, y2)
}
\arguments{
\item{data}{Matrix (x1,x2) of size n x 2}

\item{y1}{Vector of size n1}

\item{y2}{Vector of size n2}
}
\value{
\item{cdf}{Empirical cdf}
}
\description{
This function computes the empirical joint cdf evaluated at all points (y1,y2)
}
\examples{
data(simgumbel)
out=BiEmpCdf(simgumbel,c(0,1),c(-1,0,1))


}
