% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfun.R
\name{bootstrapfun}
\alias{bootstrapfun}
\title{Function to perform parametric bootstrap for goodness-of-fit tests}
\usage{
bootstrapfun(object)
}
\arguments{
\item{object}{object of class `statcvm`.}
}
\value{
\item{Sn}{Simulated value of the Cramer-von Mises statistic}

\item{Tn}{simulated value of the Kendall's statistic}

\item{Rn}{simulated value of the Spearman's statistic}

\item{parB}{Estimated parameter}
}
\description{
This function simulates Cramer-von Mises statistic and Kendall's statistic using parametric bootstrap for arbitrary data.
}
\references{
Nasri & Remillard (2023). Identifiability and inference for copula-based semiparametric models for random vectors with arbitrary marginal distributions. arXiv 2301.13408.

Nasri & Remillard (2023). Goodness-of-fit and bootstrapping for copula-based random vectors with arbitrary marginal distributions.
}
\keyword{internal}
