% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CdfInv.R
\name{CdfInv}
\alias{CdfInv}
\title{Quantile function}
\usage{
CdfInv(u, y, Fn)
}
\arguments{
\item{u}{Vector of probabilities}

\item{y}{Ordered values}

\item{Fn}{Cdf}
}
\value{
\item{x}{Vector of quantiles}
}
\description{
This function computes the inverse of the cdf of a finite distribution for a vector of probabilities.
}
\examples{
y=c(0,1,2)
Fn = c(0.5,0.85,1)
out=CdfInv(c(1:9)/10,y,Fn)


}
