% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplac.R
\name{dplac}
\alias{dplac}
\title{Density of Plackett copula}
\usage{
dplac(data, rotation = 0, par)
}
\arguments{
\item{data}{Matrix (x,y) of size n x 2}

\item{rotation}{Rotation: 0 (default value), 90, 180, or 270.}

\item{par}{Copula parameter >0}
}
\value{
\item{pdf}{Density}
}
\description{
This function computes the density of the Plackett copula with parameter par>0.
}
\examples{
dplac(c(0.5,0.8),par=3,rotation=270)


}
