% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpCdf.R
\name{EmpCdf}
\alias{EmpCdf}
\title{Empirical univariate cdf}
\usage{
EmpCdf(x)
}
\arguments{
\item{x}{Observations}
}
\value{
\item{Fx}{Empirical cdf}

\item{Fxm}{Left limit of the empirical cdf}

\item{Ix}{Indicator of atoms}data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABIAAAASCAYAAABWzo5XAAAAWElEQVR42mNgGPTAxsZmJsVqQApgmGw1yApwKcQiT7phRBuCzzCSDSHGMKINIeDNmWQlA2IigKJwIssQkHdINgxfmBBtGDEBS3KCxBc7pMQgMYE5c/AXPwAwSX4lV3pTWwAAAABJRU5ErkJggg==
}
\description{
This function computes the empirical cdf evaluated at all sample points
}
\examples{
data(simgumbel)
out=EmpCdf(simgumbel[,1])


}
