% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstDep.R
\name{EstDep}
\alias{EstDep}
\title{Kendall's tau and Spearman's rho}
\usage{
EstDep(data)
}
\arguments{
\item{data}{Matrix or data frame with 2 columns (X,Y). Can be pseudo-observations.}
}
\value{
\item{tau}{Kendall's tau}

\item{rho}{Spearman's rho}
}
\description{
This function computes Kendall's tau and Spearman's rho for arbitrary data. These are invariant by increasing mappings.
}
\examples{
data(simgumbel)
out=EstDep(simgumbel)


}
\references{
Nasri (2022). Test of serial dependence for arbitrary distributions. JMVA

Nasri & Remillard (2023). Tests of independence and randomness for arbitrary data using copula-based covariances, arXiv 2301.07267.
}
