% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnumber.R
\name{fnumber}
\alias{fnumber}
\title{Family number corresponding to VineCopula package}
\usage{
fnumber(family)
}
\arguments{
\item{family}{Copula family: "gaussian", "t", "clayton", "frank", "gumbel", "joe", "plackett'', "bb1", "bb6", "bb7","bb8".}
}
\value{
\item{fnumber}{Number}
}
\description{
Computes the  number associated with a copula family  (without rotation)
}
\examples{
fnumber("bb1")

}
\references{
Nagler et al. (2023). VineCopula: Statistical Inference of Vine Copulas, version 2.4.5.
}
