% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplac.R
\name{hplac}
\alias{hplac}
\title{Conditional distribution of Plackett copula}
\usage{
hplac(data, cond_var, rotation = 0, par)
}
\arguments{
\item{data}{Matrix (x,y) of size n x 2}

\item{cond_var}{Conditioning variable (1 or 2)}

\item{rotation}{Rotation: 0 (default value), 90, 180, or 270.}

\item{par}{Copula parameter >0}
}
\value{
\item{h}{Conditional  cdf}
}
\description{
This function computes the  conditional distribution of the Plackett copula with parameter par>0.
}
\examples{
hplac(c(0.5,0.8),1,270,3)


}
