% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoplackett.R
\name{rhoplackett}
\alias{rhoplackett}
\title{Spearman's rho for Plackett copula}
\usage{
rhoplackett(cpar, rotation = 0)
}
\arguments{
\item{cpar}{Copula parameter; can be a vector.}

\item{rotation}{Rotation: 0 (default value), 90, 180, or 270.}
}
\value{
\item{rho}{Spearman's rho}
}
\description{
Computes the theoretical Spearman's rho for Plackett copula
}
\examples{
rhoplackett(3,rotation=90)


}
\references{
Remillard (2013). Statistical Methods for Financial Engineering. CRC Press
}
