% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplac.R
\name{rplac}
\alias{rplac}
\title{Generates observations from the Plackett copula}
\usage{
rplac(n, rotation = 0, par)
}
\arguments{
\item{n}{Number of pairs to be generated}

\item{rotation}{Rotation: 0 (default value), 90, 180, or 270.}

\item{par}{Copula parameter >0}
}
\value{
\item{U}{Matrix of observations}
}
\description{
This function generates observations from a Plackett copula with parameter par>0.
}
\examples{
rplac(10,rotation=90,par=2)


}
