    \name{simgumbel}
     \docType{data}
     \alias{simgumbel}
     \title{Simulated data}
     \description{Simulated data from a Gumbel copula with parameter 2, Bernoulli margin for X1 and zero-inflated Gaussian margin for X2.}
     \usage{data(simgumbel)}
     \format{Data frame of numerical values}
     \keyword{datasets}

\examples{
data(simgumbel)
plot(simgumbel,xlab="X1", ylab="X2")
}
\keyword{datasets}

