% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taucop.R
\name{taucop}
\alias{taucop}
\title{Kendall's tau for a copula family}
\usage{
taucop(family_number, cpar, rotation = 0)
}
\arguments{
\item{family_number}{Integer from 1 to 10}

\item{cpar}{Copula parameters}

\item{rotation}{Rotation: 0 (default value), 90, 180, or 270.}
}
\value{
\item{tau}{Kendall's tau}
}
\description{
This function computes Kendall's tau for a copula family
}
\examples{
taucop(4,2,270) # Gumbel copula


}
