% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauplackett.R
\name{tauplackett}
\alias{tauplackett}
\title{Kendall's tau for Plackettfamily}
\usage{
tauplackett(cpar, rotation = 0)
}
\arguments{
\item{cpar}{Copula parameter >0}

\item{rotation}{Rotation: 0 (default value), 90, 180, or 270.}
}
\value{
\item{tau}{Kendall's tau}
}
\description{
This function computes Kendall's tau for Plackett family using numerical integration
}
\examples{
tauplackett(2,270)


}
